some manual changes
